+++ packages/loaders/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @verdaccio/loaders
 
+## 8.0.0-next-8.11
+
+### Patch Changes
+
+- Updated dependencies [f8a321f]
+  - @verdaccio/core@8.0.0-next-8.21
+
+## 8.0.0-next-8.10
+
+### Patch Changes
+
+- Updated dependencies [6a4d6dd]
+  - @verdaccio/core@8.0.0-next-8.20
+
 ## 8.0.0-next-8.9
 
 ### Patch Changes